Move Server\UserGroups to ConfigStorage namespace
